Add commenting
Added comments to the code. In part because it's the right thing to do, but also in part after discussion about the research-creation method in directed study with Rilla, Cody, and Tony. Notably the idea of technical documentation being an important component of a project. Especially if this is meant to be some kind of 'open source' it's ridiculous to not comment what the code does for someone who might want to read it.

Should probably look into javadoc or some sort of documentation generation setup? Although it's not like I'm writing a damn API...