Add slower colossus behaviour, standard snake length, snake gets longer, process notes
Added: Mostly just tweaks to get the colossus to slow down for testing purposes (now too slow) and to simplify movement code, and to make the snake actually grow on eating.

Added a process note about how 'fun' the game could be, and the idea of 'levels' with different colossi in them.

? It occurs to me that the snake getting longer in this version of the game (with each apple eaten) is actually really nice because it gives you greater 'purchase' on the colossus (easier to trap it with your longer body, though of course also easier to run into yourself). Kind of the opposite of the 'fatigue' idea, but perhaps related to the longer term idea in Shadow of getting greater grip strength.