Add basic goal scoring and reset, add commenting to Sssensible code
Most importantly added the ability to score a goal (ball goes through the goal and the snake's reset, and there's currently just a console printout of the goal)

Also went through and commented my Sssensssible code because I'm teaching programming and feel genuinely guilty about it, which is funny.

x Noted a bug in the code where body collisions aren't working. Sad!

One funny "emotion" related thing with the technical side here is that I'm acknowledging goals with classic "Goooooooooool!" statements. Somehow that made me really like the game and feel excited and happy about it. Which is a simple trick. But also goes to show just how vital the aesthetics of presentation are - even just _imaginging_ a kind of excitement associated with a goal being scored is enough to really inject a lot of spirit into the thing. Which is not quite what I would have expected? Like if I'd just written "Snake 1 scores" I don't think that would have registered with me. So it's a nice micro-example of aesthetics and emotions. Textual aesthetics in this case?