Add score display, add time display, add snakes get longer on goal, fix self-body-collision
Mostly just adding the basic niceties of a soccer game. Specifically a time counting up and a display of the two scores at the top.

Snakes get longer by a set amount each goal.

Fixed a bug that meant snakes weren't colliding with themselves.

? Snakes getting longer. Wanted to include the idea of snakes getting longer _somehow_ in the game and settled on the most basic idea that both snakes get longer with each goal scored, so there's a kind of cumulative difficulty of manoeuvring over time. It could be assymetric obviously, but I can't quite understand whether it's an advantage to be longer (more able to block) or shorter (more agile, less likely to self-collide). Also can't decide whether you would reward the goal scorer, or compensate the one scored on. Soccer itself doesn't have assymetries like that in the first place, so I kind of suspect that it's better if the growth thing is symmetric or not there, and I think it should be there. It does add something funny to play.

? Problem of ball against the side. Realised while thinking about different kinds of collisions and thinking about actual play that you would be able to push the ball against the wall and then it would become totally impossible to get it away from there. One response would be to leave it that way and demand both snakes kill themselves to reset play? The other response would be to treat it like an "out" thing if you push it to the wall and reset the ball (to the centre?). Not sure how visually/mechanically clear that would be? Another solution would be to make the goals the width of the playing area so you can still score from the wall. Currently I prefer the "out" solution, apart from its potential lack of clarity - like why _shouldn't_ you be able to push the ball to the wall? Maybe another tile indicating the "out" area or something? Actually this is starting to make sense. Although does this then suggest an open play field and no walls? That's a pretty different conception, goes away from some of the spirit of snake... this seems like a bit decision point.

? Handling of collision and ball stealing. I'm not overly satisfied with the resolution of collisions somehow, even though they're basically "fair"? One thing I'm not handling right now is the idea the ball can't go _through_ a snake's body. I probably should handle that case.

Along with that the collision with the other snake it kind of unsatisfying because it's in motion? Like it makes sense you die when you hit a snake, but the fact it just kind of glides by removes the sense of power of the death. I suppose one option would be to pause the hit snake a couple of ticks so there's more of a BAM... actually you could freeze until the dead snake respawns? Or actually maybe they flash a bit then disappear and the other snake has a free shot at the goal (unless they die themselves by mistake)? This is another big moment of decision.