Added the start of a Snake AI to guide missiles to targets
I did a really bad job of this calculations for this. Basically there's a chase() method that can cause a snake to make a move toward a target point. But right now it's just the "greedy" version, which means it will move straight down the screen and then do a diagonal waggle when it has to. I want it to be a nicer version that would approximate the diagonal line you'd draw from the start point to the target, but I ate two bagels and a donut and I can't think of how to do it. I got really stuck. I'll pick it up again later.