Added actual working code for missile snakes
Missile snakes now successfully move to the apple they are targeting and eat it. Importantly they move in a nice-looking way instead of going straight and then going into a sudden waggling motion. Jonathan helped me understand how to do that, after I spent possibly as many as 3 hours trying to figure it out for myself unsuccessfully (and was not even close by the way).

? The only musing I would add to this, I suppose, is just what a huge waste of time it was trying to fight my way through to implementing this algorithm. I really don't find it at all interesting to think about as a problem and just stuck at it because I felt like I _should_ be able to solve it. But I couldn't, and so it was to some extent just three hours down the drain. I suppose this is the kind of thing where having a collaborator or an underling of some kind to take care of this stuff would come in handy, but really I should have just asked for help faster and not been a dumbass.