Added basic code for Msss. Pacman, mini apple image, in-progress screenshot, source for level shape, tweaks
Added the base code for Msss. Pacman which means the maze is laid out and you can move in it, there are "mini apples" where you would have the normal dots, and "full apples" where you could have the power pellets, the snake wraps around on the screen when you go through the warp tunnel bit, mini apples make you get 1/25th longer (only applied when it reaches a full unit.

Also added a couple of process things: the actual Ms. Pacman map I used to draw my own (with a slight change for dimensions), and a screenshot of the game in progress.

? Most enjoyable design moment with the implementation here was thinking through the question of snake growth (as always) and the issue that if the snake grew a unit for every pellet it would be totally unwieldy. But then while I was looking at the source game and size of the pellets there, and then drawing my own, I realised they would be miniature apples and thus a _fraction_ of the value and effect of full sized apples. So the full apples are 20x20 and the mini apples are 4x4 and thus one twenty-fifth the power of the larger apples! HAH!