Added a journal entry going through my design rationale for Msss. Pacman
This involved thinking through the idea of modality in the different versions I've been making and deciding that this should be a game without modes

Also a mini-realisation about the possibility of apples in Minesweeper

And a final determination of the rules of Msss. Pacman that should allow me to finish it.