Minesssweeper: added win state, added walls around play area, added clearing text grid on game over, add showing bombs on game over
Largely added elements to do with the game over state. Specifically clearing the text grid of all the numbers so that you can successfully display a game over (or you won) text. Added the two game over states and distinguished them. Also chose to remove all unflipped tiles and reveal the locations of the bombs on game over.

? Doesn't feel like any interesting design decisions happened here. Maybe very slightly that I set the snake to 'dead' on victory so it flashes and doesn't move when you win, to avoid the situation where you have to keep steering it after the game is over, potentially dying. I could obviously have implemented a version where it's not strictly dead, just frozen, but the flashing looks kind of celebratory anyway?