Sssensssible: Added an exhausting amount of UI fixes for mobile
This was mostly a pass to increase the usability of the mobile version of Sssensssible Sssoccer by working with the idea that players play across from each other on the phone. As such I needed to implement rotatable texts in the game so I could display information to the "upside down" player and rearrange the information on the screen appropriately so that everyone can see what's happening. It wasn't exciting.

? I mean, obviously there are design implications involved in this kind of thing, particularly the idea that in a sport like soccer there's a real premium on making sure the game is "fair" in all its elements, which includes UI display. In fact there's still a small potential imbalance which is that the top goal is one tile closer to the centre than the bottom goal (this is a result of an even number of rows on the screen). It's conceivably that could be leveraged somehow? I really don't know and I don't know I can care to that level of depth.

? At the very least the information display strikes me as "fair and balanced" now which is in keeping with the spirit of the game. And frankly it feels a bit nice to be considerate about mobile play and to take it seriously with this game.