Test candidate!
This is a build that opens on the menu and is the first version I think I'd be willing to call ready for testers. (And, frankly, might be ready for release - not clear how much testing I want?)

The only niggle I have right now is that I feel like the colossus moved down onto the snake's body during a playthrough on mobile. I will look for that code.