Added press kit (no trailer), fixed text bugs in Sssensssible
Mostly just wrote the press kit and took images. Decided not to have a trailer for the game as it feels like each game is so instant in terms of meaning that it will be too spoilers-y to have any portrayal. Even the screenshots are pretty revelatory, but at least they don't give you the animation and sound so you might still play the game after seeing them?