Added process discussion (post mortem thinking, research-creation post mortem, concept post-mortem plan)
Ended up writing quite a lot here as I try to 'wrap up' the project in some kind of cohesive way that could serve as a summary of the work I've done here (to stand alongside the game itself, the repository and all its detail). Basically wanting to create two small essays that help capture the spirit of the project (as RC and as design work) that people could encounter as a first look as opposed to having to comb through the repo.